feat: add likeCount to post entity

+++ src/main/java/com/jay/sapapi/dto/post/response/PostResponseDTO.java
@@ -17,7 +17,7 @@ public class PostResponseDTO {
 
     private Long viewCount;
 
-    private int commentCount, likeCount;
+    private Long commentCount, likeCount;
 
     private String postImageUrl;
